cronet: add //testing
This is technically only used for gtest_prod.h which we could get
otherwise. However, in the current state, Chromium directly includes
gtest_prod.h and does not add a dependency on gtest.
Chromium includes are generally messy and we have applied many
downstream fixes to be able to pull in our own dependencies. We need to
revisit this topic with the Chromium team once the IC deadline has
passed.

Test: none
Change-Id: I0aabcab4cc4e2ed7015a93ace1f181e64b632e30