Introduce is_cronet_build and move process_launcher into its own target
Cronet doesn't require the setup and handling of multiple processes
(it simply spawns threads within the application process that uses
it). This wouldn't be that big of a problem, but many of the
multiprocess tooling makes use of AndroidX, which Cronet is trying not
to depend on.

To achieve that, move process_launcher files into its own target not
included in base_java (which Cronet depends on). Also, drop
IntentUtils for Cronet's builds (through the use of the new GN arg
is_cronet_built).

Test: m
Change-Id: I84f0d81f4feddce8094b6f400f6291a78c953dff