Revert "Address more Android API comments from https://r.android..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a61ce2b0-5430-4a06-9614-33eac03c4f5c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If7f172b894246dc738995e0b009c94a255f1e0e6