Setup urlconnection for Cronet MTS
This cl fixes urlconnection tests imported from chromium
and adds them to the build target.

In chromium, the test_runner runs each test on separate JVM which allowed us to run the urlconnectiontests without resetting the urlstreamfactory. This is not the case for here where the tests are run on the same JVM instance, and Factory cannot be set more than once.

Test: atest NetHttpCoverageTest
Bug: 267353182
Change-Id: Ia4b3ad5d485946bd31f0fdc1c6184d36d7238bda