Revert "Cronet: Append `mts-tethering` to native test modules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ccb2bed1-0125-4c85-bb2c-cc63362f3fb4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibb26c43e646bbbf8c4c984c2a59234dee55212ca