Revert "Cronet: remove `soong_namespace`"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2f68fbf1-3b7a-4e9c-bb17-ecf8158cc674). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1169f7c12c488a0645bd5f8a42d84f99cc8dbc41