Revert "Enable CronetUploadTest."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/85e5bac7-77a1-47ae-a9dd-8ae5ee187d6e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5b3ac72932efcf9957e9e0e18a249a026998ca90