Revert "Enable CronetLoggerTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/08d0255f-81a5-4dad-9b7a-054af1e973f1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I982219c3d84e0fabd0cba6161e2ff5525e334d47