Revert "Enable CronetLoggerTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9206a1db-1a45-4803-b735-72aa1ff73a90). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I386baf591aef78bf651ade18fecd171dcd9c288b