Revert "Cronet: Fix native data coverage for mts-tradefed"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7e23c91c-f4b0-47b7-9c61-49960a40ea34). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7b6cbd01cb49333389b33ea4041dc9b27ebfdd23