Revert "Cronet: Fix native data coverage for mts-tradefed"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7ff3f324-e2e7-4e84-91b6-b6ecdc0175b0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idcaa9ba62fc9ff9e6e2883b65fc01fe36bebbe25