Revert "Cronet: Fix native data coverage for mts-tradefed"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4282998f-e10d-47a9-8385-ce41110aaaf5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iac152c4cd346aea97f61c907b9619de49920fdf0