Revert "Cronet: Convert cc_object to cc_library_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/89453f99-aace-44ae-969d-305211e89360). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9a399eef8b14e8fcdf00621f2015374f8f61fc6d