[Cronet] Enable CronetInputStreamTest
This was failing because getMoreData of CronetHttpUrlConnection(CHUC)
was being called instead of that of MockHttpUrlConnection(MHUC).

In theory, getMoreData() of MHUC is meant to override that of
CHUC but it doesn't because the method was package private and
the packages of MHUC and CHUC are now different due to jarjaring
and one class is in the apk while the other is in the bootclasspath.

The compiler does not catch this because the package are the same
during compilation. Hence, make the method public so it can be
seen and overriden.

Test: atest NetHttpCoverageTests
Bug: 267353182
Change-Id: I2f61bcb4657b343982bcedaa2c10a4d39ed928af