Revert "Diasble memtag_globals for boringssl (inside cronet)."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/625e3f8d-13ce-4bc7-a7c5-991f518fe114). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6bf1d8e5ac10b1e4ac3a01e5c2f016159224ddd8