Revert "Diasble memtag_globals for boringssl (inside cronet)."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/11fbb3d5-2ef1-416d-9790-d23f8a713f2c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4234374a010c1be59c7ac2ec37701cf70aba5b9d