Revert "Diasble memtag_globals for boringssl (inside cronet)."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/066a40b8-0ec2-4276-9ad2-c7134ae62e04). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If649a3ab512300afde1fcb5b6df8dda911b72f58