Revert "Drop patch as that file is no longer renamed"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fd99ba66-7c52-43e9-8f1b-dcaeb68fb14c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I92d04b813ce872f932ca0ec92c69b26297f1639f