Revert "Drop patch as renaming as we now wrap instead of renaming"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fd99ba66-7c52-43e9-8f1b-dcaeb68fb14c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0b414a153b397305a777b7df34b58d6d87ff520f