Revert "Drop patch as we're no longer renaming stuff"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fd99ba66-7c52-43e9-8f1b-dcaeb68fb14c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib2839eaccea050428029730a8b1b1e3c7b789d19