Revert "Drop patch as providers have been reintroduced downstream"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fd99ba66-7c52-43e9-8f1b-dcaeb68fb14c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic6845d55a170498a779431cf4ccca2f167ede4d2