Revert "HttpEngine: make use of CronetEngine wrapping"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fd99ba66-7c52-43e9-8f1b-dcaeb68fb14c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaebb7af7f68e17c7c50c4cc7f210c6567d93d45a