Revert "Fix visibility of notifyObserversOfConnectionTypeChange"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fd99ba66-7c52-43e9-8f1b-dcaeb68fb14c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I122d7901a0e16d7fa9505c47ea72a223dfcce496