Revert "Drop patch as providers have been reintroduced downstream"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/160c9ddd-7c51-4515-a55b-6b8c3dd621bc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4616c76a17abb2d0acf2dd484306ed0ebb17a3b4