Revert "Drop patch as that file is no longer renamed"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/160c9ddd-7c51-4515-a55b-6b8c3dd621bc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5b16408ff36973f2ae3cd1103be15842027135ac