Revert "Drop patch as renaming as we now wrap instead of renaming"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/160c9ddd-7c51-4515-a55b-6b8c3dd621bc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I30e2f040a6efdb82b163c5460fdaf03b112e4d88