Revert "Fix visibility of notifyObserversOfConnectionTypeChange"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/160c9ddd-7c51-4515-a55b-6b8c3dd621bc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icc6c789d7bd1d2ea54f92b1f601d25d6dd6e13e7