Revert "Reintroduce Cronet's org.chromium.net.urlconnection APIs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3b056c13-c12d-46f2-8cc1-472835ec7559). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I498f0fa6d161ad570f7d4e4f01af334ad7b879f6