Revert "Fix visibility of notifyObserversOfConnectionTypeChange"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b77252fc-20ec-4447-adf5-c5e01e13ee8b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2b7ef1c4a9cefcd5e1e0bbf6d288abbf0170e5bd