Revert "Fix visibility of publicBuilderHttpCacheMode"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b77252fc-20ec-4447-adf5-c5e01e13ee8b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If0baf1d1754713b79150b3ded4728f6a4bb66b91