Revert "Drop patch as custom library loader has been reintroduce..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b77252fc-20ec-4447-adf5-c5e01e13ee8b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie6b482695a00edaa71a38496cfbd70d9bbec7968