Revert "Drop patch as renaming as we now wrap instead of renaming"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b77252fc-20ec-4447-adf5-c5e01e13ee8b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6ac346c9aa38199155dc1a5857c729293ea46bab