Revert "HttpEngine: make use of CronetEngine wrapping"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e767d3bd-583f-4738-9ed7-cba624c4bc90). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id2f10afa1c5a62e8b9507f8bd6897952c884e213