Revert "Drop patch as we moved HeaderBlock back into UrlResponseInfo"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e767d3bd-583f-4738-9ed7-cba624c4bc90). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib7d2050cf5d0c766f5c5debec2fc9606c85a824d