Revert "Drop patch as that file is no longer renamed"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e767d3bd-583f-4738-9ed7-cba624c4bc90). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1d246ddcb32ac9558477096132d2a34232516b88