Revert "Add UploadDataSinkWrapper"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cd3ebf89-25b4-4431-81d4-3c1440cd3629). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia88177ba0d85b20d9d1644b4b26cfb3c48fd2d68