Revert "net_tests: don't depended on cronet"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cd3ebf89-25b4-4431-81d4-3c1440cd3629). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If8d2e206cd5b56de314d3d0bbae8377b30fc8c2b