Revert "Merge remote-tracking branch 'origin/master' into upstre..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cd3ebf89-25b4-4431-81d4-3c1440cd3629). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6bfa5dc2e75c37e2bde3836a28f7dfac2644af4b