Revert "HttpEngine: make use of CronetEngine wrapping"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b2447ae7-5b20-4ce6-9f42-72f4a0462c81). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I62506e305bce63bcf3ef5ee3e4eff93d5636d8be