Revert "Drop patch as we're no longer renaming stuff"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b2447ae7-5b20-4ce6-9f42-72f4a0462c81). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I90bc01955d4ce51905203ff55ace2a2bd55918c7