Revert "net_tests: don't depended on cronet"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/de4a8706-9303-48be-aa0f-ccb9fc3b9966). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8009729ee5392240b544ad1b1c719c708c059cab