Revert "Reintroduce Cronet's org.chromium.net.urlconnection APIs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/de4a8706-9303-48be-aa0f-ccb9fc3b9966). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If9cec48c1554205ad558c3de925bdf86d4af279e