Revert "Add BidirectionalStreamBuilderWrapper"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/de4a8706-9303-48be-aa0f-ccb9fc3b9966). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4d75d7056bc12030469deb7559238c61a3b05c08