Revert "Drop downstream experimental option translation logic ch..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/de4a8706-9303-48be-aa0f-ccb9fc3b9966). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ief0af8bebef310e39d92d61cedcbd8b67c487f45