Revert "Add UrlRequestBuilderWrapper"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b0a64faf-522f-4358-b806-c24eb60ba260). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0cd8580409010a1e69265897ed6423e53f71a813