Revert "Add BidirectionalStreamBuilderWrapper"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b0a64faf-522f-4358-b806-c24eb60ba260). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia88d4231fcd28a74fbad52b401d74a3b8017a836