Revert "Merge remote-tracking branch 'origin/master' into upstre..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d63587f2-45c3-4387-97ea-e8337a6e18fe). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I99b542b2c4416ab280335c524cfacdc7e5f3e7a6