Revert "Reintroduce Cronet's org.chromium.net.urlconnection APIs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d63587f2-45c3-4387-97ea-e8337a6e18fe). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iec20381fefb69a772d6d0f5093d64c4f8bf5c8c3