Revert "Drop patch as renaming as we now wrap instead of renaming"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c9e29d89-72b0-46e4-b2c5-8d691e929e09). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5a52beb16fb8deedc43b72a203a20ef8f6ce8e62