Revert "Add CronetEngineWrapper"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f9774301-f2b8-468a-9044-f681e1d06a60). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iced151f2f43f78795cf0599361d9cc6e2db74254