Revert "Add BidirectionalStreamCallbackWrapper"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f9774301-f2b8-468a-9044-f681e1d06a60). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8f7aabcf284bc8f68ca5b6df470327b86f354d4d