Revert "Drop patch as custom library loader has been reintroduce..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1bbe29c7-8ca5-4739-bb14-0fafaaa25352). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I50f927da31231b9d170b7fe2c9adf11277424dcb