Revert "Fix visibility of publicBuilderHttpCacheMode"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1bbe29c7-8ca5-4739-bb14-0fafaaa25352). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9d616d61c9c36b5829fd0ce7b78b1516ba593e6a