Revert "Drop patch as renaming as we now wrap instead of renaming"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1bbe29c7-8ca5-4739-bb14-0fafaaa25352). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5572a2679c7e775ae5932cffa2193f3ca33ca815