Revert "net_tests: don't depended on cronet"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a4050c76-beee-4075-b4cc-c3628d8c7d90). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I891901b51090cb20c838eb53b679d216ac9a0b36