Revert "Fix BidirectionalStream getters"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a4050c76-beee-4075-b4cc-c3628d8c7d90). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I051a9180f3a71c3f2b255f5db9db07742ed6a534