Revert "Drop downstream experimental option translation logic ch..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a37e6735-59bb-47f0-a1fa-5d7f6311b0ef). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic5bc6a1a04b9eddb03367794098145f30e7866e0