Revert "Cronet's javatest: go back to targeting org.chromium.net..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a37e6735-59bb-47f0-a1fa-5d7f6311b0ef). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6bd6d36d60a3ced4ab2e684704ec39806f6b0e01