Revert "Add UrlResponseInfoWrapper"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7b388b38-eee9-44d6-97ba-3f59d30e5b06). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaf0aebddf85b6d72ce6b0a1421cea6026f7e6ebe