Revert "Drop patch as custom library loader has been reintroduce..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6e1863b5-de52-46f6-a35c-ed199bff3d99). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1ae7b88fba3f20e6e7673fc9efa397bf93a85471