Revert "Drop patch as we're no longer renaming stuff"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6e1863b5-de52-46f6-a35c-ed199bff3d99). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6ece6405f5ec8565ce2a5913e6328825e2b10f5d