Revert "net_tests: don't depended on cronet"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/43dbe900-6286-4a98-baad-ca57082972c0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib4ef3d6a41791ffbb937b6960d02c08f2ba0a7d6