Revert "Merge remote-tracking branch 'origin/master' into upstre..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/43dbe900-6286-4a98-baad-ca57082972c0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I83ab1c906263dfa765ea44c85a89bd15afc74aee