Revert "Add UploadDataSinkWrapper"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/43dbe900-6286-4a98-baad-ca57082972c0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4c234f77a734e23ab370a9a973df1ae199df9939