Revert "Drop downstream experimental option translation logic ch..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a6c6a93f-1a78-4d4c-bd37-c3f471c84fad). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I25bacac64fb0eaf4ea02225add9f1c1a0978a8b1