Revert "net_tests: don't depended on cronet"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a6c6a93f-1a78-4d4c-bd37-c3f471c84fad). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I021d012db8ab2735723fc28fda2f3c4ddf9da3a9