Revert "Add UrlRequestStatusListenerWrapper"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/36fe37bd-d2ff-4628-9928-82abba6a6d48). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic6e44f4a0369abefb6bde65a99d50dedee0bd265