Revert "Merge remote-tracking branch 'origin/master' into upstre..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/36fe37bd-d2ff-4628-9928-82abba6a6d48). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I34dc030f8666c419e1147476e52d12c3c77251e4