Revert "Merge remote-tracking branch 'origin/master' into upstre..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ef64620e-7ae0-4c93-9c3f-227f0f47a0a4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie985b499c4e378d0e4e1af7274d63de344cfd8af