Revert "Reintroduce Cronet's org.chromium.net.urlconnection APIs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ef64620e-7ae0-4c93-9c3f-227f0f47a0a4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8a62b2a9dcc060d6e47a26355ae469a6976a4334