Implement missing methods not existing in Chromium Cronet's API
Those getters/setters were implemented as per the API Council's requirement to ship Cronet to Android U as
seen here (aosp/2646733) and here (aosp/2647252). When we introduced Cronet Wrapper API, it was implemented
upstream (Chromium) so it matched the Cronet's API that exist upstream. However, HttpEngine API is not identical
to Cronet's API, which has caused an error.

Bug: 314275954
Change-Id: I1ceb61a98cfe30329afb1234bf9f0ff54dba9afb