Revert "Sandbox cronet genrules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3bc0ee8b-fa31-4022-b3b8-55f4d77b0dd5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2dfcc3b0fcfc1838a9bcab3079e4b2c74a07169b