Revert "Sandbox cronet genrules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/479ce58c-a0a4-47e0-aff3-cbc5ac75b16a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I93dca8394f6f17752759d74b937a7bd6437bd0fa