Revert "Sandbox cronet genrules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/806fa68f-09d7-42eb-9b69-b15d5c9eb079). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5a2ec7fad14926da18b1e4ae89c7445f58f9bd01