Revert "Sandbox cronet genrules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f6c513db-fb74-4b8b-9907-272f42e39197). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id1bc394025221da550ab73f7a4321a2189036bf6