Revert "Sandbox cronet genrules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e1a0c5f5-5998-464d-989d-c1ba249686d6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie759e0a5cd7d6b3df9c9e8ed6d221768eadd6bb0