Revert "Sandbox cronet genrules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ce877a63-89e8-4e50-9487-cf8d4211fddf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6edbc3962395bdddbcc751e4e0e68aa970bd8c34