Revert "Sandbox cronet genrules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9c7428b8-5da8-4381-913f-97bc91765c77). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I189739069e385e4c40877375876f9b78d4d17220