Revert "Sandbox cronet genrules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/533dc982-659c-46a4-8719-91dcd2327abc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia364186b2dda4f43169e5538ada56d8599279c97