Revert "Sandbox cronet genrules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f37a4c36-18d7-4e01-82d0-dea20f6c7746). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id9e12c8501435da5469e95fcd44c4d0268bec0c5