Revert "Sandbox cronet genrules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8dc0a772-773b-410f-9cd3-6b651f1080b9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If23f4b265e0c32d714be8f1a2cd64521f0664e01