Revert "Sandbox cronet genrules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5e90302c-40c1-4b22-a8d0-798ddf62efbf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0ead3a469fd8377e21629ed5150c387b24d9a813