Revert "Sandbox cronet genrules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ec2ddb52-9f4a-4c5f-97d1-b636d2b993eb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If7c0fbd7b31962ae153a5bbd2444e4ad069cdf0c