Revert "Sandbox cronet genrules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1bab647f-7a54-4ceb-b058-1ce176b79441). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9b7d5c3a7dfac16df2fd791bd5028988f67e2799