Revert "Sandbox cronet genrules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/83d8c5c5-7f69-4d2a-b36f-34e8fa2f8ac3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8554cb683fa5372f76dde27471254767624454ca