Revert "Sandbox cronet genrules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3172fb82-9d6d-4593-9222-868301a024ea). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I39bce7fecbf0415eda3dd8dd6f752f0b07a3b79b