Revert "Sandbox cronet genrules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/422e7d5f-557b-433c-bc24-36ec01f6ea4c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3fc231821aa8e188d6f1eeff91d1dba2bb227137