Implement proper JNI registration handling
This is a continuation of the previous CL.

* For the java side, Add all possible java files that were found in `generate_jni_header`
targets to the java genrule srcs. This will ensure that the native method will exist and
can be called correctly.
* For the cc side, only add the java files that were found in `generate_jni_header`
targets AND are part of the transitive dependency of the GN target.

Note: Look into removing the cc side of JNI registration since it does not provide extra correctness

Bug: 304217500
Change-Id: Ia302616090e68815e507986d416f129290585eca