Re-add the rest of the tests
The tests were filtered to remove tests that requires HTTP2 test server, the filter
is no longer required as we imported netty and the tests are now passing.

Add a filegroup for cronet testing data, this was not correctly exported
to the device. Look for the changes in AndroidTest.xml within Connectivity/cronet/tests/mts
module to see the changes

Bug: 319629150
Bug: 304217500
Change-Id: Ifc096f4197c547202f266ea843d1c5bf70353bcd