Revert "Remove base divergence"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5e6be325-95b9-4a6e-8cfc-e6d2ac50af98). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia93b454a6c6ce963b54605878b6457532d164218