Revert "Remove base divergence"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3fd3283a-2ad1-40cd-b45f-10336ea69598). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7f2a6b3e71bf5d153dead0cf44c039f42e45b919