Revert "Remove base divergence"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8e02430d-6402-4904-94fe-c24d74c61beb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9e35fd247128cf1f26e1b67caa780b1bf002601a