Revert "Remove base divergence"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f9ac56e9-05e2-4a4a-bde9-0b99d57aa122). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic04d350a07fe55bbdc7e13814f8a2cd0d1213ba8