Revert "Revert "Revert "Cronet 121_0_6167_71: Merge remote-track..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fc78f391-3093-49a0-ad8b-9bf21cc533b5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3243229b15c67cbe569cd0ccf6dc65fbbacc3847