Try to run cronet tests with a custom classloader
This should never be merged, it is simply a trial to re-enable
the failing test because some code exercise cronet on the bootclasspath
and the rest of the code use statically linked cronet in the APK.

This is one-approach to solve this problem by loading everything from
the statically linked cronet in the APK, this guarantees that all the
tests will use the same shared library. However, this comes with a cost -
which is a classloader can lead to "hard to fix" bugs due to collisions
between different classloader that may happen, lemme give an example in this CL

While trying to implement this, I had a problem where no tests were executed at all.
This was because the junit annotation `@Test` was loaded twice, it was loaded from
the app class loader(the original class loader in the APK) but when the tests themself
are loaded by the custom class loader, which means `@Test` is also loaded one more time
, so when Junit does getAnnotation().equals(Test.class), this fails because the annotation
stored in `getAnnotation` is from custom class loader while Test.class is from the original
class loader.

The workaround for the above problem is to never load anything that does not begin with
`android.net.connectivity` package name, this ensures that all of our third party libraries
will be loaded once, and for Cronet code, there is no way to load Cronet code except from the test class.

Another problem that I run into was HttpEngine, HttpEngine API was not statically linked in the test APK
which means that the classes had to be loaded from the bootclasspath, and HttpEngine wrapper has a dependency
on CronetEngine which resulted in an error because CronetEngine class now got loaded from two different classloaders.

The solution to this problem was to add HttpEngine API to the test APK as seen in the other CL.

Test: atest NetHttpCoverageTest
Change-Id: I98388ed1e0bec11097fca53366e5e69a84f0fc1c