Remove build timestamp verification in net unittests
Previously, net_unittests_tester would fail after 70 days of stale
Cronet import which would result in a P0 bug that can fixed by a simple
one-liner CL. We've tried to fix this in multiple ways mentioned in
b/294801110 and none of them has worked.

In summary, we tried setting the timestamp in the future but that has
failed as the branch condition uses `magnitude` which is an absolute value.

We also tried making the timestamp value dynamic which resulted in
non-reproducable builds and that's a mainline requirement to have
reproducable builds.

This change will introduce divergence which is acceptable because this
file is rarely touched and any merge conflict should be straightforward
to fix, this is not the optimal scenario and we will most likely Remove
this condition upstream if the tests proved to work well without it.

Bug: 294801110
Test: TH
Change-Id: I63e0cd03e765255cd3d1705c7ce040f3db542416