Move disabled net tests to TEST_MAPPING
Those tests are only disabled in AOSP and not on Chromium and they
would cause divergence on import. In order to avoid that, I moved those
tests to TEST_MAPPING instead, the benefit of doing that is that the
flaky tests can still run on postsubmit while not blocking the
presubmits.

Bug: 304217500
Test: TH
Change-Id: Id2c03b7ed2e77fc835c2ee8f7114059b57069bff