Cronet 123.0.6312.40: Update access modifiers to fix tests
Some tests calls production code which is not public but under the same
package.
NetHttpCoverageTests only jarjar the production code. So test can no
longer access these production code.

Possible solutions would be 1) jarjar both tests and production code,
2) Update jarjar_excludes, and 3) make production code public.

1st solution would require more amount of work.
2nd solution needs to add many classes to jarjar_excludes since non
public class that is accessed from the class in jarjar_excludes also
need to be added to the jarjar_excludes.

So this CL did the 3rd option.
This change should be upstreamed.

Test: build
Bug: 342969117
Change-Id: I0a0e91335d6aa108f87b88a415bbbf74cfc635fb