Revert "Update the build timestamp."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7f13a46a-3c15-4023-abd4-8f62cfd91a0c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibbb002032f4d1fb7ef7efc0ad9e7ed06a99a6794
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com