Revert "Update the build timestamp."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/40bb5f22-d2c8-4442-acec-62586afe88fd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0ff7f78a4efcfe6393bf27aad32a6cc93cd89bc3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com