Add support for statically generated build scripts output
Rust targets that exists in the build_script_output.json file will
now have their rustflags extended with the generated data from build
scripts. It's assumed that the output of build script will always be
passed to the rust toolchain through rustflags. this might not be
correct in the future but for now, this is working and should keep working.

Tweaked the additional_args logic to accommodate for extending, previously
it would overwrite the entire field, this changes make it extend the
current value with the extra field if it is a set.

build_script_outputs.json file was generated in Chromium v126.0.6423.0
and copied manually to AOSP.

See go/rust-assumptions-for-gn2bp for more information about the build
scripts and their output.

Bug: 346568402
Change-Id: Ia08591eb86c5325c83c8f263375f83f2be821e52