Add support for rust_library and rust_proc_macro
* Rust targets are always groups and an assumption was made in GN2BP that
groups will always carry configs and never real targets, this CL changes
that so that group dependencies will be visited and materialized in the
final Android.bp if they need to.

* Few fixes to action_foreach as sources and tool_files were no
propagated properly, also the dependencies of action_foreach were ignored
due to a bug where merging arch-specific variant were not correct.

* GN2BP assumes that if a target X supports arch-variant Y then direct
child dependencies of X MUST support the same arch-variant Y. this is
an invalid assumption as an android-only action can depend on a host
only action. In order to work around this problem, I decided to add
another hack (if statement) which basically merge that "host" variant
into the common variant.

Bug: 346568402
Test: None
Change-Id: I4388438cae6a633e976cc55260358d457977000e